Treating signer indices as u64

+++ mithril-core/benches/merkle_tree_create.rs
@@ -29,10 +29,10 @@ pub fn merkle_tree_verify_benchmark(c: &mut Criterion) {
     let mut rng = OsRng::default();
     let ps = gen_keys(32);
     let mt = MerkleTree::<H>::create(&ps);
-    let i = rng.next_u64() as usize % 32;
+    let i = rng.next_u64() % 32;
     let path = mt.get_path(i);
     c.bench_function("Merkle Tree verify 2**5", |b| {
-        b.iter(|| mt.to_commitment().check(&ps[i], i, &path))
+        b.iter(|| mt.to_commitment().check(&ps[i as usize], i, &path))
     });
 }
