+++ mithril-core/examples/dynamic_stake.rs
@@ -178,7 +178,7 @@ fn main() {
     println!("+------------------------+");
 }
 
-fn local_reg(ids: &[(usize, u64)], pks: &[MspPk]) -> ClosedKeyReg<H> {
+fn local_reg(ids: &[(u64, u64)], pks: &[MspPk]) -> ClosedKeyReg<H> {
     let mut local_keyreg = KeyReg::new(ids);
     // todo: maybe its cleaner to have a `StmPublic` instance that covers the "shareable"
     // data, such as the public key, stake and id.
